agregando librerias de jQuery  y aprendieneo PHP

+++ js/main.js
@@ -1,44 +1,47 @@
 (function() {
     "use strict";
+    if (document.getElementById('slogan')) {
+        //  typewriter
 
-    // typewriter
+        var app = document.getElementById('slogan');
 
-    var app = document.getElementById('slogan');
+        var typewriter = new Typewriter(app, {
+            loop: true,
+            delay: 75,
+        });
 
-    var typewriter = new Typewriter(app, {
-        loop: true,
-        delay: 75,
-    });
-
-    typewriter
-        .pauseFor(2000)
-        .typeString('La mejor conferencia WEB del 2021')
-        .pauseFor(300)
-        .deleteAll()
-        .pauseFor(100)
-        .typeString('Aprenderas los mejores trucos de <strong style="color: yellow;">JS</strong> ')
-        .pauseFor(300)
-        .typeString('<strong></br>Hecho con <span style="color: #fe4918;">❤️</span> en San Gil</strong>')
-        .pauseFor(1000)
-        .start();
-    // fin typewritter
-
+        typewriter
+            .pauseFor(2000)
+            .typeString('La mejor conferencia WEB del 2021')
+            .pauseFor(300)
+            .deleteAll()
+            .pauseFor(100)
+            .typeString('Aprenderas los mejores trucos de <strong style="color: yellow;">JS</strong> ')
+            .pauseFor(300)
+            .typeString('<strong></br>Hecho con <span style="color: #fe4918;">❤️</span> en San Gil</strong>')
+            .pauseFor(1000)
+            .start();
+        //  fin typewritter
+    }
 
     var regalo = document.getElementById('regalo');
 
     document.addEventListener('DOMContentLoaded', function() {
 
-        //          mapa
-        var map = L.map('mapa').setView([6.556432, -73.129478], 17);
+        if (document.getElementById('mapa')) {
+            //          mapa
+            var map = L.map('mapa').setView([6.556432, -73.129478], 17);
 
-        L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-            attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-        }).addTo(map);
+            L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
+                attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
+            }).addTo(map);
 
-        L.marker([6.556432, -73.129478]).addTo(map)
-            .bindTooltip('Hogar.<br> Dulce hogar')
-            .openTooltip();
-        //       fin mapa
+            L.marker([6.556432, -73.129478]).addTo(map)
+                .bindTooltip('Hogar.<br> Dulce hogar')
+                .openTooltip();
+            //       fin mapa
+        }
+
 
         //       Datos del usuario
         var nombre = document.getElementById("nombre");
@@ -99,39 +102,42 @@
                 // console.log('eror= ', this.value.indexOf("@"));
             }
         }
+        if (document.getElementById('calcular')) {
+            function calcularMontos(event) {
+                event.preventDefault();
+                if (regalo.value === '') {
+                    alert('ELIJE UN REGALO PARA CONTINUAR');
+                    regalo.focus();
+                } else {
+                    var boletosDia = parseInt(pase_dia.value, 10) || 0,
+                        boleto2Dias = parseInt(pase_dosDias.value, 10) || 0,
+                        boletoCompleto = parseInt(pase_completo.value, 10) || 0,
+                        cantEtiquetas = parseInt(etiquetas.value, 10) || 0,
+                        cantCamisas = parseInt(camisas.value, 10) || 0;
 
-        function calcularMontos(event) {
-            event.preventDefault();
-            if (regalo.value === '') {
-                alert('ELIJE UN REGALO PARA CONTINUAR');
-                regalo.focus();
-            } else {
-                var boletosDia = parseInt(pase_dia.value, 10) || 0,
-                    boleto2Dias = parseInt(pase_dosDias.value, 10) || 0,
-                    boletoCompleto = parseInt(pase_completo.value, 10) || 0,
-                    cantEtiquetas = parseInt(etiquetas.value, 10) || 0,
-                    cantCamisas = parseInt(camisas.value, 10) || 0;
+                    var totalPagar = (boletosDia * 30) + (boleto2Dias * 45) + (boletoCompleto * 50) + ((cantCamisas * 10) * 0.93) + (cantEtiquetas * 2);
 
-                var totalPagar = (boletosDia * 30) + (boleto2Dias * 45) + (boletoCompleto * 50) + ((cantCamisas * 10) * 0.93) + (cantEtiquetas * 2);
+                    var listadoProductos = [];
+                    if (boletosDia >= 1) { listadoProductos.push(boletosDia + ' pases por dias'); }
+                    if (boleto2Dias >= 1) { listadoProductos.push(boleto2Dias + ' pases por dos dias'); }
+                    if (boletoCompleto >= 1) { listadoProductos.push(boletoCompleto + ' pases completos'); }
 
-                var listadoProductos = [];
-                if (boletosDia >= 1) { listadoProductos.push(boletosDia + ' pases por dias'); }
-                if (boleto2Dias >= 1) { listadoProductos.push(boleto2Dias + ' pases por dos dias'); }
-                if (boletoCompleto >= 1) { listadoProductos.push(boletoCompleto + ' pases completos'); }
+                    if (cantCamisas >= 1) { listadoProductos.push(cantCamisas + ' numero de camisas'); }
+                    if (cantEtiquetas >= 1) { listadoProductos.push(cantEtiquetas + ' paquetes de etiquetas'); }
 
-                if (cantCamisas >= 1) { listadoProductos.push(cantCamisas + ' numero de camisas'); }
-                if (cantEtiquetas >= 1) { listadoProductos.push(cantEtiquetas + ' paquetes de etiquetas'); }
+                    lista_productos.style.display = "block";
+                    lista_productos.innerHTML = '';
+                    for (var i = 0; i < listadoProductos.length; i++) {
+                        lista_productos.innerHTML += listadoProductos[i] + '</br>';
+                    }
+                    //funcion Intl. es un constructor me devuelve las cifras en unidades cientos y millares
+                    suma_total.innerHTML = '$ ' + (new Intl.NumberFormat("en-US").format(totalPagar));
 
-                lista_productos.style.display = "block";
-                lista_productos.innerHTML = '';
-                for (var i = 0; i < listadoProductos.length; i++) {
-                    lista_productos.innerHTML += listadoProductos[i] + '</br>';
                 }
-                //funcion Intl. es un constructor me devuelve las cifras en unidades cientos y millares
-                suma_total.innerHTML = '$ ' + (new Intl.NumberFormat("en-US").format(totalPagar));
-
             }
         }
+
+
 
         function mostrarDia() {
             var boletosDia = parseInt(pase_dia.value, 10) || 0,
@@ -183,12 +189,20 @@ $(function() {
 
         return false;
     });
+    // activar animaciones en una pagina al dar scroll con waypoints
+    var resumenLista = jQuery('.resumen-evento'); //seleccion de la seccion con jquery o $
+    if (resumenLista.length > 0) { // cuando aparezca la seccion en la ventana el length sera 1
+        $('.resumen-evento').waypoint(function() { // con waypoints se ejecuta dentro de la funcion las animaciones
+            //animaciones para los numeros
+            $('.resumen-evento li:nth-child(1) p').animateNumber({ number: 6 }, 1200);
+            $('.resumen-evento li:nth-child(2) p').animateNumber({ number: 15 }, 1200);
+            $('.resumen-evento li:nth-child(3) p').animateNumber({ number: 3 }, 1500);
+            $('.resumen-evento li:nth-child(4) p').animateNumber({ number: 9 }, 1200);
+        }, {
+            offset: '60%' // la animacion se ejecuta cuando vaya por un 60% de la pantalla
+        });
+    }
 
-    //animaciones para los numeros
-    $('.resumen-evento li:nth-child(1) p').animateNumber({ number: 6 }, 1200);
-    $('.resumen-evento li:nth-child(2) p').animateNumber({ number: 15 }, 1200);
-    $('.resumen-evento li:nth-child(3) p').animateNumber({ number: 3 }, 1500);
-    $('.resumen-evento li:nth-child(4) p').animateNumber({ number: 9 }, 1200);
 
     // plugin de countdown o cuentaregresiva
 
@@ -199,7 +213,7 @@ $(function() {
         $('#segundos').html(e.strftime('%S'));
     });
 
-    //scroll
+    //  SCROLL
     var windosheight = $(window).height(); //altura de la ventana del navegador
     var barraAltura = $('.barra').innerHeight(); //altura de la barra
 
@@ -219,6 +233,12 @@ $(function() {
             $('body').css({ 'margin-top': '0px' })
         }
 
+    });
+
+    // Menu Responsive
+
+    $('.menu-movil').on('click', function() {
+        $('.navegacion-principal').slideToggle();
     })
 
 });
